+++ coil_module.f90
@@ -9,9 +9,9 @@ module coil_module
    integer :: num_periods=4, is_mirrored=0, main_winding=1
    integer :: num_main_coils, skip_value=1
    integer :: main_size, aux_size, main_count, aux_count, taper_size, aux_flag
-   real, allocatable, dimension(:,:,:) :: coil_main, coil_aux
-   real, allocatable, dimension(:) :: main_current, aux_current, taper
-   real, dimension(:), allocatable :: aux_percent
+   double precision, allocatable, dimension(:,:,:) :: coil_main, coil_aux
+   double precision, allocatable, dimension(:) :: main_current, aux_current, taper
+   double precision, dimension(:), allocatable :: aux_percent
    integer, allocatable, dimension(:) :: main_points, aux_points
    integer :: num_aux_coils = 0
    integer :: main_current_repeat = 0
